+++ src/pages/homePage/index.tsx
@@ -1,0 +1,11 @@
+import { useEffect } from 'react';
+
+function HomePage() {
+  useEffect(() => {
+    console.log("a", )
+  }, [])
+
+  return <div>home page ne</div>
+}
+
+export default HomePage;