+++ src/types/http/header.ts
@@ -1,0 +1,9 @@
+export const HTTP_HEADER_KEY = {
+  AUTHORIZATION: 'authorization',
+  CONTENT_TYPE: 'Content-Type',
+}
+
+export const HTTP_HEADER_VALUE = {
+  APPLICATION_JSON: 'application/json',
+  URLENCODED: 'application/x-www-form-urlencoded'
+}