+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## CHANGELOG
+
+## 08/04/2022
+- [core] update env for application && add structure for api
+
 ## 04/04/2022
 - [feature] 
   - setup eslint + prettier